+++ app/src/main/java/com/pbs/expenseApp/ui/components/AppIcon.kt
@@ -1,0 +1,23 @@
+package com.pbs.expenseApp.ui.components
+
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+
+@Composable
+fun AppIcon(
+    modifier: Modifier = Modifier,
+    imageVector: ImageVector,
+    tint: Color = MaterialTheme.colorScheme.primary,
+    contentDescription: String
+) {
+    Icon(
+        imageVector = imageVector,
+        contentDescription = contentDescription,
+        tint = tint,
+        modifier = modifier
+    )
+}