+++ app/src/main/java/com/pbs/expenseApp/ui/components/AppNavigationBar.kt
@@ -1,11 +1,9 @@
 package com.pbs.expenseApp.ui.components
 
-import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.NavigationBar
 import androidx.compose.material3.NavigationBarItem
 import androidx.compose.material3.NavigationBarItemDefaults
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.res.stringResource
@@ -14,10 +12,11 @@ import com.pbs.expenseApp.appstate.AppState
 import com.pbs.expenseApp.navigation.BottomBarRoutes
 
 @Composable
-fun BottomBarNavigation(appState: AppState) {
+fun AppNavigationBar(appState: AppState) {
     NavigationBar(
         containerColor = MaterialTheme.colorScheme.primaryContainer,
     ) {
+        // TODO move to model view
         val navigationBarItems = listOf(
             BottomBarRoutes.ADD_EXPENSE,
             BottomBarRoutes.RESUME,
@@ -31,13 +30,10 @@ fun BottomBarNavigation(appState: AppState) {
             NavigationBarItem(
                 selected = item.route == currentRoute,
                 label = {
-                    Text(
-                        text = title,
-                        color = MaterialTheme.colorScheme.onPrimaryContainer
-                    )
+                    AppText(id = item.title, color = MaterialTheme.colorScheme.onPrimaryContainer)
                 },
                 icon = {
-                    Icon(
+                    AppIcon(
                         imageVector = item.icon,
                         contentDescription = "$title icon",
                         tint = MaterialTheme.colorScheme.onPrimaryContainer