+++ app/src/main/java/com/pbs/expenseApp/ui/components/AppRow.kt
@@ -1,0 +1,22 @@
+package com.pbs.expenseApp.ui.components
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.RowScope
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+
+@Composable
+fun AppRow(
+    modifier: Modifier = Modifier,
+    rowContent: @Composable() RowScope.() -> Unit
+) {
+    Row(
+        modifier = modifier,
+        verticalAlignment = Alignment.CenterVertically,
+        horizontalArrangement = Arrangement.Center
+    ) {
+        rowContent()
+    }
+}