+++ app/src/main/java/com/pbs/expenseApp/ui/composables/MyMonthlySavingText.kt
@@ -1,0 +1,17 @@
+package com.pbs.expenseApp.ui.composables
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.text.font.FontWeight
+import com.pbs.expenseApp.R
+import com.pbs.expenseApp.ui.components.AppText
+
+@Composable
+fun MyMonthlySavingText(monthlySavings: Int) {
+    AppText(id = R.string.add_expense_monthly_savings_1)
+    Text(
+        text = " $monthlySavings€ ",
+        fontWeight = FontWeight.Bold
+    )
+    AppText(id = R.string.add_expense_monthly_savings_2)
+}